+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This page lists the main changes made to Myokit in each release.
 
 ## Unreleased
 - Added
+  - [#872](https://github.com/MichaelClerx/myokit/pull/872) The `myokit.Simulation` class can now be initialised using a precompiled backend (see the documentation for the ``path`` argument).
 - Changed
   - [#872](https://github.com/MichaelClerx/myokit/pull/872) `Model.solvable_order` now returns consistently ordered equations, regardless of the order components and variables were created/parsed in.
   - [#872](https://github.com/MichaelClerx/myokit/pull/872) `myokit.tools.rmtree` now takes an optional argument `silent` that can be set to `True` to ignore any exceptions.